requirements.txt
Flask==2.1.1
psycopg2-binary==2.8.6
joblib==0.17.0
pandas==1.4.2
sklearn==0.0
openai==0.18.0
replicate==0.0.1a14
flask-cors
authlib==1.0.1
python-dotenv
Flask-Login
Flask-Bcrypt
Flask-SQLAlchemy~=2.5.1
authlib
flask-praetorian~=1.3.0
stripe
